fix: removed default value in OsuVersion property fix: added handling for double value in EmitIniKeysValue fix: added support for undocumented legacy values in event type fix: switched offset and MillisecondPerBeat to double as that is what they should be (contrary to what the wiki says) fix: improved the GetKeyPairs method, now correctly ignoring comments test: improved testing for GetKeyPairs method test: added generalSectionTest

+++ Library/TimingPoint.cs
@@ -23,12 +23,12 @@ namespace OsuLoader {
         /// <summary>
         /// Offset of the track timing (if inherithed is the distance from the last non inherithed timing point)
         /// </summary>
-        public int Offset { get; set; }
+        public double Offset { get; set; }
 
         /// <summary>
         /// Time for each beat, if inherited, negative inverse slider velocity multiplier instead.
         /// </summary>
-        public float MilliSecondPerBeat { get; set; }
+        public double MilliSecondPerBeat { get; set; }
 
         /// <summary>
         /// Beats per measure, ignored if inherited
